+++ changelog.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
+## [Unreleased] - SC
+### Added
+- Entity Class
+- Task class that extends entity and has setter methods
+
 ## [Unreleased] - YR 
 ### Update
 - Size, group, and status in form to add a new todo item